refactor(cli-tests): split e2e suites and trim redundant cases
