+++ packages/cli/src/drift.e2e.test.ts
@@ -131,10 +131,6 @@ async function createFixture(database: string): Promise<E2EFixture> {
 describe('@chx/cli drift depth env e2e', () => {
   const liveEnv = getRequiredEnv()
 
-  test('validates required env variables are present', () => {
-    expect(() => getRequiredEnv()).not.toThrow()
-  })
-
   test(
     'detects manual drift and exposes reason counts via drift/check JSON',
     async () => {
